feat(mcp): extract CLI option parsing into cli-options-lib
Move MCP CLI argument parsing and help text helpers into a focused lib
module with a thin re-export wrapper and targeted tests.
